requirements.txt
allure-pytest==2.15.0
allure-python-commons==2.15.0
attrs==25.3.0
black==25.1.0
certifi==2025.8.3
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.2.1
distlib==0.4.0
execnet==2.1.1
filelock==3.19.1
h11==0.16.0
identify==2.6.13
idna==3.10
iniconfig==2.1.0
isort==6.0.1
MouseInfo==0.1.3
mypy_extensions==1.1.0
nodeenv==1.9.1
outcome==1.3.0.post0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
pre_commit==4.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==1.0.9
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-Quartz==11.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==8.4.1
pytest-xdist==3.8.0
python-dotenv==1.1.1
pytweening==1.2.0
PyYAML==6.0.2
requests==2.32.5
rubicon-objc==0.5.2
ruff==0.12.11
selenium==4.35.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
virtualenv==20.34.0
websocket-client==1.8.0
wsproto==1.2.0
